refactoring and wwith to api (dev)
